requirements.txt
click
colorama
halo
PTable
pylint
pyyaml
azure-mgmt-keyvault
haikunator
keyring
azure.keyvault